+++ cpp/optimizers/LazyGreedyOptimizer.cpp
@@ -18,12 +18,12 @@ bool LazyGreedyOptimizer::equals(double val1, double val2, double eps) {
   }
 }
 
-std::vector<std::pair<ll, float>> LazyGreedyOptimizer::maximize(
+std::vector<std::pair<ll, double>> LazyGreedyOptimizer::maximize(
     SetFunction &f_obj, ll budget, bool stopIfZeroGain = false,
     bool stopIfNegativeGain = false, bool verbose = false) {
     //TODO: take care of handling equal guys later
 	//TODO: take care of different sizes of each items - becomes a candidate only if best and within budget, cost sensitive selection
-    std::vector<std::pair<ll, float>> greedyVector;
+    std::vector<std::pair<ll, double>> greedyVector;
     greedyVector.reserve(budget);
     std::unordered_set<ll> greedySet;
     greedySet.reserve(budget);
@@ -46,18 +46,18 @@ std::vector<std::pair<ll, float>> LazyGreedyOptimizer::maximize(
     }
     f_obj.clearMemoization();
     // initialize priority queue:
-    std::priority_queue<std::pair<float, ll>> maxHeap;
+    std::priority_queue<std::pair<double, ll>> maxHeap;
     // for each element in the ground set
     for (auto elem : groundSet) {
         // store <elem, marginalGainWithMemoization(greedySet, elem)> in
         // priority-queue (max-heap)
-        maxHeap.push(std::pair<float, ll>(
+        maxHeap.push(std::pair<double, ll>(
             f_obj.marginalGainWithMemoization(greedySet, elem), elem));
     }
     while (rem_budget > 0) {
-        std::pair<float, ll> currentMax = maxHeap.top();
+        std::pair<double, ll> currentMax = maxHeap.top();
         maxHeap.pop();
-        float newMaxBound =
+        double newMaxBound =
             f_obj.marginalGainWithMemoization(greedySet, currentMax.second);
         if (newMaxBound > maxHeap.top().first) {
             // add currentMax.first to greedy set after checking stop conditions
@@ -71,7 +71,7 @@ std::vector<std::pair<ll, float>> LazyGreedyOptimizer::maximize(
                         .second);  // greedily insert the best datapoint index
                                    // of current iteration of while loop
                 greedyVector.push_back(
-                    std::pair<ll, float>(currentMax.second, newMaxBound));
+                    std::pair<ll, double>(currentMax.second, newMaxBound));
                 rem_budget -= 1;
                 if (verbose) {
                     std::cout << "Added element " << currentMax.second
@@ -84,7 +84,7 @@ std::vector<std::pair<ll, float>> LazyGreedyOptimizer::maximize(
                 }
             }
         } else {
-            maxHeap.push(std::pair<float, ll>(newMaxBound, currentMax.second));
+            maxHeap.push(std::pair<double, ll>(newMaxBound, currentMax.second));
         }
     }
 		return greedyVector;
