+++ cpp/optimizers/NaiveGreedyOptimizer.h
@@ -7,7 +7,7 @@ class NaiveGreedyOptimizer
 {
     public:
     NaiveGreedyOptimizer();
-    std::vector<std::pair<ll, float>> maximize(SetFunction &f_obj, ll budget, bool stopIfZeroGain, bool stopIfNegativeGain, bool verbose);
+    std::vector<std::pair<ll, double>> maximize(SetFunction &f_obj, ll budget, bool stopIfZeroGain, bool stopIfNegativeGain, bool verbose);
     bool equals(double val1, double val2, double eps);
 };
 #endif
