+++ cpp/optimizers/StochasticGreedyOptimizer.h
@@ -7,7 +7,7 @@ class StochasticGreedyOptimizer
 {
     public:
     StochasticGreedyOptimizer();
-    std::vector<std::pair<ll, float>> maximize(SetFunction &f_obj, ll budget, bool stopIfZeroGain, bool stopIfNegativeGain, float epsilon, bool verbose);
+    std::vector<std::pair<ll, double>> maximize(SetFunction &f_obj, ll budget, bool stopIfZeroGain, bool stopIfNegativeGain, float epsilon, bool verbose);
     bool equals(double val1, double val2, double eps);
 };
 #endif
